+++ packages/calcite-components/src/components/block/block.e2e.ts
@@ -75,6 +75,10 @@ describe("calcite-block", () => {
         propertyName: "sortDisabled",
         defaultValue: false,
       },
+      {
+        propertyName: "scale",
+        defaultValue: "m",
+      },
     ]);
   });
 
@@ -111,6 +115,10 @@ describe("calcite-block", () => {
       {
         propertyName: "sortHandleOpen",
         value: true,
+      },
+      {
+        propertyName: "scale",
+        value: "m",
       },
     ]);
   });
@@ -523,6 +531,16 @@ describe("calcite-block", () => {
           "--calcite-block-border-color": {
             targetProp: "borderColor",
           },
+          "--calcite-block-content-space": [
+            {
+              shadowSelector: `section.${CSS.content}`,
+              targetProp: "paddingBlock",
+            },
+            {
+              shadowSelector: `section.${CSS.content}`,
+              targetProp: "paddingInline",
+            },
+          ],
           "--calcite-block-header-background-color": {
             shadowSelector: `.${CSS.toggle}`,
             targetProp: "backgroundColor",