+++ packages/calcite-components/src/components/dropdown/dropdown.tsx
@@ -637,7 +637,7 @@ export class Dropdown
             aria-controls={IDS.menu(guid)}
             ariaExpanded={open}
             ariaHasPopup="menu"
-            name={SLOTS.dropdownTrigger}
+            name={SLOTS.trigger}
             onSlotChange={this.updateTriggers}
           />
         </div>